Rag bot hdfc credit card guide/requirements.txt
streamlit
numpy
requests
faiss-cpu
sentence-transformers
